feat: 7주차 변경사항 추가
