+++ src/components/footer/Footer.js
@@ -10,10 +10,10 @@ function Footer() {
     <div className="Footer">
     <div className="content">
         <div className="box">
-          <div className="item01">@codeit - 2023</div>
+          <span className="item01">@codeit - 2023</span>
           <div className="item02">
-          <div>Privacy Policy</div>
-          <div>FAQ</div>
+          <span>Privacy Policy</span>
+          <span>FAQ</span>
           </div>
           <div className="item03">
           <FooterSns value={facebookImage} />
