+++ src/components/main/emptyFile.js
@@ -1,7 +1,7 @@
 export function EmptyFile() {
   return (
-  <div className="EmptyFile">
+  <span className="EmptyFile">
     저장된 링크가 없습니다. 
-  </div>
+  </span>
   )
 }
