+++ src/components/modal/modal.js
@@ -5,8 +5,8 @@ const customImgStyle = {
   width: "24px",
   height: "24px",
   position: "absolute",
-  top: "4%",
-  right: "4%",
+  left: "350px",
+  top: "8px",
 }
 
 const customModalStyles = {
@@ -20,13 +20,14 @@ const customModalStyles = {
     left: "0",
   },
   content: {
-    width: "300px",
-    height: "300px",
+    width: "45vh",
+    height: "45vh",
     overflow: "visible",
     zIndex: "150",
     position: "absolute",
-    top: "20%",
-    left: "35%",
+    top: "50%",
+    left: "50%",
+    transform: "translate(-50%, -50%)",
     borderRadius: "10px",
     boxShadow: "2px 2px 2px rgba(0, 0, 0, 0.25)",
     backgroundColor: "white",
