--- src/components/modal/shareModal.css
@@ -1,90 +1,0 @@
-.ShareModal {
-  display: flex;
-  padding: 32px 40px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 24px;
-}
-.ShareModal .titles {
-  display: flex;
-  width: 280px;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 8px;
-}
-.ShareModal .titles .title {
-  color: var(--Linkbrary-gray100, #373740);
-  font-family: Pretendard;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-}
-.ShareModal .titles .subTitle {
-  color: var(--Linkbrary-gray60, #9FA6B2);
-  text-align: center;
-  font-family: Pretendard;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 22px; /* 157.143% */
-}
-.ShareModal .share {
-  display: flex;
-  align-items: flex-start;
-  gap: 32px;
-}
-.ShareModal .share .link {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 10px;
-}
-.ShareModal .share .kakao .dumy {
-  width: 2rem;
-  height: 2rem;
-  display: flex;
-  padding: 12px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  border-radius: 37.333px;
-  background: #FEE500;
-}
-.ShareModal .share .kakao .dumy > img {
-  width: 1.2rem;
-  height: 1.2rem; 
-}
-.ShareModal .share .facebook .dumy {
-  width: 2rem;
-  height: 2rem;
-  display: flex;
-  padding: 12px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  border-radius: 37.333px;
-  background: #1877F2;
-}
-.ShareModal .share .facebook .dumy > img {
-  width: 1.2rem;
-  height: 1.2rem; 
-}
-.ShareModal .share .copy .dumy {
-  width: 2rem;
-  height: 2rem;
-  display: flex;
-  padding: 12px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  border-radius: 37.333px;
-  background: rgba(157, 157, 157, 0.04);
-}
-.ShareModal .share .copy .dumy > img {
-  width: 1.2rem;
-  height: 1.2rem; 
-}
